DailySiteDistributor/requirements.txt
# MySQL Database Connection
pymysql==1.1.0

# Environment Variables Management
python-dotenv==1.0.0

# Optional: AWS SDK
boto3==1.34.34